[chore] cleaned up files, moved getEvent function
